devops-lab-platform/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
docker>=7.1.0
pyyaml==6.0.1
aiofiles==23.2.1
python-multipart==0.0.9
